requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
pillow
opencv-python
scikit-image